Move default config to header file, set it as default parameter
