+++ control/gimbal_controller.hpp
@@ -57,6 +57,11 @@ typedef struct
 	attitude_command_t   		attitude_output_config;  			///< Initial output commands
 } gimbal_controller_conf_t;
 
+/**
+ * \brief Default configuration structure
+ */
+static inline gimbal_controller_conf_t gimbal_controller_default_config();
+
 
 class Gimbal_controller
 {
@@ -67,7 +72,7 @@ public:
 	 * \param   controller          Pointer to data structure
 	 * \param   config              Pointer to configuration
 	 */
-	Gimbal_controller(Servo& servo_4,Servo& servo_5, const gimbal_controller_conf_t config);
+	Gimbal_controller(Servo& servo_4,Servo& servo_5, const gimbal_controller_conf_t config = gimbal_controller_default_config() );
 
 
 	/**
@@ -100,6 +105,36 @@ private:
 	attitude_command_t			attitude_output_;			///< Output to PWM (output)
 	Servo&						servo_pitch_;				///< Gimbal pitch servo
 	Servo&						servo_yaw_;					///< Gimbal yaw servo
+};
+
+
+static inline gimbal_controller_conf_t gimbal_controller_default_config()
+{
+    gimbal_controller_conf_t conf = {};
+
+    float min_max_angle = 45.0f;
+
+    //init desired attitude command
+    conf.attitude_command_desired_config.rpy[0] = 0.0f;
+    conf.attitude_command_desired_config.rpy[1] = 0.0f;
+    conf.attitude_command_desired_config.rpy[2] = 0.0f;
+
+    //init output commands
+	conf.attitude_output_config.rpy[0] = 0.0f;
+	conf.attitude_output_config.rpy[1] = 0.0f;
+	conf.attitude_output_config.rpy[2] = 0.0f;
+
+	//init MIN range of allowed attitude command
+	conf.attitude_command_range_config[0].rpy[0] = -min_max_angle;
+	conf.attitude_command_range_config[0].rpy[1] = -min_max_angle;
+	conf.attitude_command_range_config[0].rpy[2] = -min_max_angle;
+
+	//init MAX range of allowed attitude command
+	conf.attitude_command_range_config[1].rpy[0] = min_max_angle;
+	conf.attitude_command_range_config[1].rpy[1] = min_max_angle;
+	conf.attitude_command_range_config[1].rpy[2] = min_max_angle;
+
+    return conf;
 };
 
 #endif /* GIMBAL_CONTROLLER_H_ */
